fixing errors
